move modulelifecycle to module-operator
