+++ module-operator/controllers/modulelifecycle/finalizer.go
