+++ module-operator/controllers/modulelifecycle/finalizer_test.go
