+++ module-operator/controllers/modulelifecycle/manager.go
